Refactor Sample Verifier tests
